refactor(domain/enums/LotteryGrade): merge `Grade` into `LotteryGrade`
